起動時にDBを読み込む処理をリファクタリングした

+++ X4_DataExporterWPF/Entity/Common.cs
@@ -15,7 +15,7 @@
         /// <summary>
         /// 値
         /// </summary>
-        public int Value { get; }
+        public long Value { get; }
         #endregion
 
 
@@ -24,7 +24,7 @@
         /// </summary>
         /// <param name="item">項目名</param>
         /// <param name="value">値</param>
-        public Common(string item, int value)
+        public Common(string item, long value)
         {
             Item = item;
             Value = value;
